+++ CHANGELOG.rst
@@ -1,5 +1,6 @@
 100.2.6-dev (Unreleased)
 --------------------
+* Added batched security reindex for BDR Collection role changes
 
 100.2.5 (2026-05-05)
 --------------------